chore: moved sortedQuickstartComponents to designated file
